🔧 Apply code-review findings for the cache generation guard
Close a residual hang in the new race tests: `waitUntilEntered` records an
issue and *returns* on expiry, so when the GET never reached the transport —
the exact regression the suite guards — the mutation went on to dequeue the
still-gated slot and park with nobody left to open the gate. `.timeLimit`
cannot break that, because `FetchGate.wait()` suspends on a continuation
cancellation does not resume. Both tests now re-check `enteredCount` and bail
out, opening the gate first.

Bring the three sibling descriptions of the cache contract back in line with
the corrected one. `CacheConfiguration`, `TMDbConfiguration.cache` and the
getting-started article all still said invalidation happens on "POST or
DELETE", omitting PUT and the state-changing GET, and correcting only the
how-to widened that divergence.

Reword the in-flight sentence: "discarded rather than cached" could be read as
the caller's response being dropped, when it is still returned — only the
cache write is refused.

Move `FetchGate` into its own `TestUtils/FetchGate.swift`. It is a
general-purpose primitive and now has consumers beyond the configuration
tests, so a file named after an unrelated mock is the wrong home. Pure move,
no signature or visibility change.
