+++ Sources/TMDb/Networking/CacheHTTPClient.swift
@@ -105,8 +105,9 @@ private actor ResponseCache {
 /// `URLCache`.
 ///
 /// Invalidation covers reads that were already in flight: a response fetched
-/// before a mutation landed carries pre-mutation state, so it is discarded
-/// rather than cached. Each read captures the cache's generation when it looks
+/// before a mutation landed carries pre-mutation state, so it is still returned
+/// to its caller but is not written to the cache. Each read captures the
+/// cache's generation when it looks
 /// up, and a mutation bumps that generation — see
 /// `ResponseCache.setResponse(_:forKey:ifGeneration:)`.
 ///
