+++ Sources/TMDb/TMDb.docc/HowTos/CachingResponses.md
@@ -79,8 +79,9 @@ is cleared when the process ends), caches `GET` requests only, bypasses
 user-specific requests (see below), and invalidates the entire cache after any
 successful `POST`, `PUT` or `DELETE`, or a state-changing `GET` such as clearing
 a list. Invalidation also covers reads already in flight: a response fetched
-before the mutation landed describes pre-mutation state, so it is discarded
-rather than cached.
+before the mutation landed describes pre-mutation state, so it is still returned
+to the caller that asked for it, but is not written to the cache — the next read
+goes back to the network.
 
 Reach for it when you want a predictable, fixed TTL regardless of TMDb's headers,
 or when you need a cache on a platform where `URLCache` is unavailable.
